pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "trojan" 
version = "0.1.0"
description = "TBD"
requires-python = ">=3.8" 

dependencies = [
    "openai>=1.0.0",        
    "python-dotenv>=1.0.0", 
    "pandas>=2.3.0",
    "matplotlib>=3.8.0",
    "tqdm>=4.65.0",
]
[tool.setuptools.packages.find]
exclude = ["data*", "results*"]